# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agent Orcha API",
        "version": "0.0.3",
        "description": "API for orchestrating agents, workflows, knowledge stores, functions, MCP servers, and LLMs."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/functions/{name}": {
            "get": {
                "summary": "Get Function Details",
                "tags": ["Functions"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Function details with parameter schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "name": { "type": "string" },
                                        "description": { "type": "string" },
                                        "parameters": { "type": "object" }
                                    }
                                }
                            }
                        }
                    },
                    "404": { "description": "Function not found" }
                }
            }
        },
        "/api/functions/{name}/call": {
            "post": {
                "summary": "Call Function",
                "tags": ["Functions"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "arguments": { "type": "object" }
                                },
                                "required": ["arguments"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Function result",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    },
                    "404": { "description": "Function not found" },
                    "500": { "description": "Internal server error" }
                }
            }
        },
...
        "/api/files/read": {
            "get": {
                "summary": "Read File",
                "description": "Read file contents by path.",
                "tags": ["Files"],
                "parameters": [
                    {
                        "name": "path",
                        "in": "query",
                        "required": true,
                        "schema": { "type": "string" },
                        "description": "Relative file path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File contents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "content": { "type": "string" },
                                        "path": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "400": { "description": "Invalid path" },
                    "404": { "description": "File not found" }
                }
            }
        },
        "/api/files/write": {
            "put": {
                "summary": "Write File",
                "description": "Write content to a file. Triggers hot-reload for config files.",
                "tags": ["Files"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "path": { "type": "string", "description": "Relative file path" },
                                    "content": { "type": "string", "description": "File content" }
                                },
                                "required": ["path", "content"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Write result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" }
                                    }
                                }
                            }
                        }
                    },
                    "400": { "description": "Invalid path" },
                    "500": { "description": "Internal server error" }
                }
            }
        },
...
    }
}